Fixed rare bug in optimal accuracy alignment. When model is configured in global mode it was possible for the optimal accuracy parsetree to include emissions from the local end (EL) state, even though they should be impossible. This causes a problem when converting the resulting parsetree into an alignment because the PP annotation for the EL emissions is undefined.
This occurs if a 0.0 PP for those EL emissions is 'optimal' (results
in the maximum sum of PPs for all emitted residues, even when 0.0 is
attributed to the EL emissions). One might argue this means they
should be allowed, but doing that is a bad idea because it violates
too many assumptions made by various parts of the alignment machinery.

This occurred only rarely when aligning random sequences. I've never 
observed it on a CM emitted sequence or on a hit found by
cmsearch/cmscan. 

The fix for this brought in in this revision is to explicitly check
for and disallow transitions to EL that imply EL emissions when local
ends are off. This was added to cm_OptAccAlign() and cm_OptAccAlignHB()
in cm_dpalign.c, and cm_TrOptAccAlign() and cm_TrOptAccAlignHB() in
cm_dpalign_trunc.c.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4053 db3e70e4-12e1-0310-b23f-81a6e3d09ba7