Revising new cmalign behavior introduced in previous revision (r4054). The failover to HMM banded non-truncated alignment from HMM banded truncated alignment now only occurs in case of 'error type 2' (described in r4054 log) where HMM bands disallow all valid parsetrees. Failover is no longer allowed in case of 'error type 1', which occurs if required size of truncated HMM band matrices exceed maximum.
The motivation for this change is that error type 1 sometimes occurs
for valid (homologous) target sequences for large models and I've
decided it's better to inform the user they should be allowing larger
matrices (with --mxsize) than force standard alignment in these cases
which will almost certainly result in poor alignments.  I'm
particularly concerned about SSU and LSU rRNA alignment, where
sequences are often truncated and the required matrices for extreme
(but valid) input sequences can exceed the 1Gb limit (especially for
LSU). I think users trying to align these large RNAs should be
encouraged to increase the max size with --mxsize, which is now what
cmalign will do once it is used to try to align one of these extreme
examples. With the failover behavior introduced in r4054 (that I'm
reverting here), some of these sequences would be aligned poorly in
standard mode, unbeknownst to the user.  


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4056 db3e70e4-12e1-0310-b23f-81a6e3d09ba7