cmsearch.c, dev_testsuite.sqc: Input sequence files can no longer be an alignment. This change was necessary because the machinery for reading sequences isn't tolerant of interleaved sequences in the input file because long sequences are not read all at once, but rather in overlapping windows. This change mirrors a similar one made to nhmmer in HMMER svn rev 3938. The testsuite now checks that a Stockholm alignment causes a cmsearch failure.
cmscan.c: 
Enforced a maximum length of 10 Mb for query sequences. This means
cmscan should be usable for essentially all prokaryotic genomes, but
not for the vast majority of eukaryotic ones. For euk genomes, users
will just be forced to use cmsearch. This is okay because cmscan is
really meant for short sequences, e.g. from a metagenomics study. 

The reason a maximum sequence length is important is that cmscan reads
in full sequences instead of reading in overlapping windows like
cmsearch does. Not reading windows significantly simplifies the code
but means that we have to limit the maximum size so we're not reading
in Gb long sequences all at once which could potentially require too
much memory.

cm_pipeline.c:
The --max and --nohmm options to cmsearch and cmscan will now turn off
hmm-only scans, forcing the CM to be used for all models, even ones
with zero basepairs. If the user is setting --max or --nohmm they
can't be worried about running time, so turning off HMM only scans
shouldn't be a problem. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4059 db3e70e4-12e1-0310-b23f-81a6e3d09ba7