cmalign.c, cmcalibrate.c, cmscan.c, cmsearch.c: The four programs that use threads now always print the number of processors being used in the output header, even if --cpu is not used. (cmalign only prints a header if user redirects alignment with -o.)
The motivation for this is that the determination of thread number was
non-intuitive to me and so I think it may also be non-intuitive to a
significant number of users. X threads are always used on an X-core
machine, even if some of those threads are being used by other
processes. This can be undesirable. For example, if I submit many jobs
to the cluster, 8 jobs may be put onto the same node because each node
has 8 CPUs. Each of those jobs will be restricted to 3 Gb of RAM by
default, because our cluster divides up the 24 Gb or RAM per node
evenly and gives 3 Gb to any single-CPU on the node. However, all 8
jobs will fire up 8 threads because they each recogize the machine
they are running on has 8 CPUs. This can cause a problem if these jobs
are memory intensive. For example, cmalign limits the summed total
size of its DP matrices to 1 Gb. Given 3 Gb of total RAM, this would
if we have 1 worker thread (which I assumed would be the case), but is
not fine if we have 8 threads because we'll potentially want 8 Gb of
RAM.

When I ran into this case, I had to modify cmalign to output the
number of threads it was using before I could make sense of what was
going on. If a user runs into the same issue, they'll be hard-pressed
to make sense of it unless they know the number of threads being used
- this is why the number of threads used is now always printed. I'll
also make a note of this issue in the user guide and man pages, but
that will only be helpful to users who actually read the
documentation.

cmbuild.c: fixed off-by-one in alignment index in error message if an
	   alignment has no SS_cons annotation. 

mpisupport.c: fixed sizeof() function for a CM_ALIDISPLAY, it was not
	      counting one integer.

cm_pipeline.c: Silenced compiler warning regarding printing an int
	       with PRId64.

Still working on tutorial. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4060 db3e70e4-12e1-0310-b23f-81a6e3d09ba7