cmsearch.c, cmscan.c, cmalign.c, sse_cmsearch.c: small changes to silence icc compiler warnings.
itest_brute.c: modified sampling of random CM parameters to ensure
that the begin probability into state 6 of the ROOT->MATL->MATL->END CM is
never 1.0, if it is then a length 2 sequence has no valid parsetree
(because the MATL_IL prior to the END gets detached to remove
ambiguity with ROOT_IR). If there's no valid parsetree for a length 2
sequence then the paranoid internal checks in the CYK dp
implementations fail and itest_brute fails. This only occured with
Intel compilers, for an unknown reason. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4084 db3e70e4-12e1-0310-b23f-81a6e3d09ba7