Added creation of a symlink to make hmmer's testsuite pass a top-level 'make check' when hmmer/easel does not exist (e.g. in a src distribution) and the build directory is not the same directory as the src directory.
hmmer's testsuite expects that hmmer/easel exists because it uses
miniapps in hmmer/easel/miniapps/ and also the BLOSUM62 file in
easel/formats/. I had a workraound for this (creating symlink to
infernal/easel if infernal/hmmer/easel does not exist), but this would
only work when the builddir and the srcdir were the same dir -- using
a builddir other than the srcdir left hmmer's testsuite unable to find
easel/formats/BLOSUM62 (because that's not in the builddir, but rather
the srcdir). So now I check for this case and create two symlinks to
easel if necessary.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4126 db3e70e4-12e1-0310-b23f-81a6e3d09ba7