Fixing problem with 'make clean' and 'make distclean' in a released distribution.
Previously, a released distribution did not include the
documentation/userguide/ directory, because building the guide
requires the executable programs 'rman', 'pdflatex', 'bibtex' to be in
the user's path and Sean doesn't want people trying to build the
guide, having it fail mysteriously with a bunch of errors and then
reporting bugs to us.

Now, a released distribution will include the documentation/userguide/
directory, but a 'make' in that directory will first check to see if
'rman', 'pdflatex' and 'bibtex exist in the user's path using the
'command -v' command. As far as I can tell using 'command' is the
safest way to do this, although I could change this to running each
command with the --version option and checking if the return status is
in error or not. 

I will update the 00CHECKLIST in the trunk soon so the next release
will include the documentation/userguide/ directory. 

configure.ac changes because previously we had to ensure
documentation/userguide/ existed before trying to create it's makefile
with AC_CONFIG, now we don't have to check because it will always
exist. 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4154 db3e70e4-12e1-0310-b23f-81a6e3d09ba7