Fixed and logged bug i28 in 1.1 release branch. This is the first bug discovered in the 1.1rc1 release. It's logged in BUGTRAX and tested for in the testsuite with the i28* files.
The bug is that the function to clone CP9Bands_t objects
(hmmband.c:cp9_CloneBands()) was not cloning all
parameters. Importantly the {J,L,R,T}valid parameters relevant to
truncated search/alignment were not being cloned. This couldn't affect
the final alignment but did cause a failure in rare cases. In
particular, it caused a possible downstread problem in truncated
pipeline passes if a single envelope contained multiple hits and the
required matrix exceeded the maximum allowed size. It is rare to have
multiple hits in a single envelope in truncated passes because they
are restricted to only the first or final W residues in a sequence. In
these rare cases the pipeline fails and exits because the matrix
should never exceed the maximum size because we've already ensured it
does not with the bands prior to cloning them. The reason it does
exceed the max size with the cloned bands is that we failed to clone
the {J,L,R,T}valid parameters. These parameters dictate what type of
truncated alignments are possible, so not copying them has the effect
that the cloned bands imply a bigger requisite matrix then the
original template bands.

The bug manifests itself once in the ENSEMBLE release 47 rhesus
macaque genome when searching with an SRP model, which finds on the
ballpark of 1 million hits (Alus). This is the only case I've ever
observed so it is what is tested in the testsuite.

More details on how the bug was discovered are in:
~nawrockie/notebook/12_0708_inf_stadler_paper_repro/00LOG.

infernal.h: modified a single comment line



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4174 db3e70e4-12e1-0310-b23f-81a6e3d09ba7