Fixed and logged bug i29, which was present in infernal 1.1rc1 (but will have been fixed in 1.1rc2, coming soon). Added test script that gets run in testsuite.sqc with 'make check'.
Bug was that rarely with version 1.0 some CM files get printed with 0
filter threshold points by cmcalibrate. Then in 1.1rc1 if we try to
use cmconvert on such a file we get an error. As of 1.0.2, I don't
think it's possible to calibrate a CM and get 0 filter threshold
points. This bug occurs for 1 out of the 1973 Rfam 10.1 files and was
reported by Jen Daub when she was trying to convert them. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4186 db3e70e4-12e1-0310-b23f-81a6e3d09ba7