Adding handling for zero-length sequences in cmsearch. Previously a zero length sequence would cause the program to fail and output an error message in serial and threaded modes. Now it simply skips these sequences and includes them in the 'number of sequences', but they, obviously, don't increase residue count at all. This commit actually only fixes the problem for serial mode. A soon-to-be-made Easel commit (likely r794) will fix the problem for threaded mode.
BUGTRAX, testsuite.sqc, dev_testsuite.sqc: added bug i30 which is the
failure on zero length sequences bug. testsuite and dev_testsuite now
run a script that checks if this bug exists or not.

configure.ac: fixed lack of handling of --enable-gsl, which was
deleted recently in r4179. configure now sets LIBGSL variable, which
is used by easel/miniapps/Makefile to "" instead of not setting it at
all, which was a problem.

itest7-trunc.pl: wasn't cleaning up a tmp file, now it does.
bug-i11.pl:      fixed a comment that referred to i6 instead of i11.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4196 db3e70e4-12e1-0310-b23f-81a6e3d09ba7