cmsearch can now handle gzipped sequence files in serial or threaded mode. Previously it would fail with an error message when given a gzipped file.
cmsearch needs to do a first pass read through the full file before it
begins a search. In the first pass, the lengths of each sequence are
recorded and the total database size is used to define the filter
thresholds that should be used. The file is then rewound for the next
pass, during which the actual search is performed. Gzipp'ed files
can't be rewound in Easel, which is why they were not previously
supported. Now, cmsearch implements a poor man's rewind of a gzipped
file by closing it and reopening it after the first pass. This was
suggested by Zasha Weinberg who requested the ability to read gzipp'ed
files. 

Reading of non-gzipped files should be completely unchanged.

It is certainly possible there's some reason this is a bad idea that I
just don't know of right now. If I learn of such a reason I'll revert
to the old behavior.

MPI cmsearch still can't handle gzipped files because it requires a
SSI index on the fasta file, and an SSI index cannot be created for a
gzipp'ed file.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4199 db3e70e4-12e1-0310-b23f-81a6e3d09ba7