This is a significant post-1.1rc1 release commit in the 1.1 release branch. All changes are necessary to enable a --glist <f> option to cmscan which reads a list of CM names from file <f> and runs those models in glocal mode instead of local.
This motivation here is for Rfam, to eliminate the need for a
rfam_scan.pl script that wraps cmscan. The problem this new option
addresses is that some of the Rfam 11 models (40 out of 2208) have had
their GA, TC, NC bit score thresholds determined for searches in
glocal model, with all others being determined in local mode. This
means to do a proper 'cmscan --cut_ga' you'd really have to run two
separate cmscan calls, one with the 40 models in glocal mode and the
other with the remaining 2168 in local mode, and then merge and resort
the output. This merging and resorting would likely be done by a
rfam_scan.pl script. The --glist option allows this to be done in a
single cmscan call, and should obviate the need for a rfam_scan.pl
script altogether.  

The --glist option has been added to the cmscan man page, and is
tested by the itest8-glist.pl script run by the testsuite.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4208 db3e70e4-12e1-0310-b23f-81a6e3d09ba7