Modified how cmalign places insert and EL columns ('.' and '~' RF columns, respectively) in a rare case when both an IL and an EL are used to emit at least one residue after the same consensus column in an alignment. Previously, IL inserts would occur 5' of the EL inserts. Now, the order is reversed: EL inserts occur 5' of IL inserts.
This only changes behavior for the following rare case: when a MATP
node is immediately followed by an END node (i.e. a
zero-consensus-length hairpin loop). When MATP->END, a MATP_IL can
emit to the same position as an EL. However, for MATL->END (the only
other x->END possibility), the MATL_IL is unreachable (to deal with an
ambiguity in the CM grammar) and an IR state necessarily emits at the
same position as the EL. Since ELs were introduced in Infernal, they
have always emitted 5' of IR states, so the convention of EL 5' of
inserts was already enforced for the vast majority of cases.

Now that ELs are *always* 5' of inserts, we can easily update
esl-alimerge to properly handle the merging of alignments that use EL
states. This is important for large alignments, e.g. SSU rRNA where it
is desirable to parallelize a large cmalign job into many smaller jobs
and then merge the resulting alignments together into one. Without
this new convention of ELs 5' of inserts, it would be possible to
update esl-alimerge to do this, but it would need to parse the SS_cons
of the alignment to determine whether ELs should be 5' or 3' of
inserts, and it would require the user to specify that an Infernal or
at least an RNA alignment was being input. I think enforcing the EL 5'
of insert convention is better as it significantly reduces the
required amount of code. The esl-alimerge miniapp will be updated
shortly. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4271 db3e70e4-12e1-0310-b23f-81a6e3d09ba7