Fixed and logged bug i33 reported by Zasha Weinberg in the 1.1 release branch. With 'cmsearch -A foo.stk foo.cm foo.fa', foo.stk can be corrupt (most commonly with invalid PP lines).
This only occurs for a small but significant fraction (7% in Rfam
11.0) of CM files created with Infernal v1.0 and converted to 1.1
format with 1.1's 'cmconvert' program. The problem lies in
cm_alidisplay.c::cm_alidisplay_BackConvert() which uses
HandModelmaker() to construct a CM guide tree from a msa->ss_cons.
The resulting guide tree does not always match the 1.0-built CM guide
tree in the CM file because the HandModelmaker() function changed
between versions 1.0.2 and 1.1 (to fix bugs i19 and i20). When the bug
occurs, cm_alidisplay_BackConvert() gets a guide tree that doesn't
correspond to the CM and tries to use it to convert an aligned
sequence to a parsetree in Transmogrify() which, of course, does not
work properly. Surprisingly, this never causes a segmentation fault
and most commonly manifests itself by outputting corrupt PP lines in
the -A output alignment file.

The fix is to construct the guide tree that Transmogrify() requires
directly from the CM data structure instead of the msa->ss_cons, which
will work if the CM was built with v1.0 or v1.1 cmbuild.

cmalign.c: removed unused variable introduces in recent revision.
itest3-msa-integrity.pl: typo fix


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4274 db3e70e4-12e1-0310-b23f-81a6e3d09ba7