Reverting some changes from recent revisions 4226 and 4231 that modified the CM file output format by changing the 'dbsize' value used for CM exponential tail stats from a long to a double. The values are now again printed as longs, as they were in 1.1rc1.
I originally changed to doubles when I fixed bug i31, because dbsize
values > 2Gb were problematic on 32-bit machines. However, changing
the output format affected reading of binary files and meant that
1.1rc1 press'd files could not be read by future versions (1.1rc2 and
beyond), and vice versa. I realized this was a bad thing and so I'm
reverting in this revision. The affected functions now check that the
dbsize value is less than 2Gb before printing, and fail if it is
not. This is okay because the maximum the value will ever be is 160Mb,
because that's the maximum value for -L <x> in cmcalibrate.

Also changed max <x> for -L <x> in cmcalibrate, previously it was 100
Mb. I increased it to 160 Mb, which is 100-fold higher than the
default value of 1.6 Mb. <x> is the total length of random sequence
searched during calibration.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4292 db3e70e4-12e1-0310-b23f-81a6e3d09ba7