Fixing bug i34 in Infernal 1.1 branch. *Should* be final commit before 1.1rc2 release. This bug concerns the rare case (1 occurence in all Rfam 10.1 models) when a MATP node is immediately followed by an END node, which is the only way a IL and EL can emit after the same consensus position. As of a recent revision (r4272), I *thought* a new convention of always placing EL columns before (5') of insert columns in output alignments was 100% implemented, but I hadn't updated the insert rejustification code in cm_parsetree.c::Parsetrees2Alignment() to reflect this new convention. This revision fixes that oversight.
Bug i34 in BUGTRAX explains this further, and bug-i34.pl in the
testsuite tests for this bug. 

Also, cmalign.c has been slightly simplified by removing code meant to
allow an upper limit of 100K seqs or 100M residues to be aligned in
interleaved or non-Stockholm output alignments. This was having no
effect, as esl_sqio_ReadBlock() enforces that only up to 1 megabyte
(1024*1024 residues) of sequence can be read in a block, so that was
the actual upper limit. The code now reflects this. (I considered
changing the Easel function to allow a caller-defined max number of
residues, but I'm trying to make the minimal amount of changes
immediately prior the Infernal 1.1rc2 release, lest I break something
and don't notice it until post-release.)


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4336 db3e70e4-12e1-0310-b23f-81a6e3d09ba7