Fixed and logged bug i35 in Infernal 1.1 release branch. This bug commonly caused failures for a specific type of MPI cmcalibrate job: when the model was large (W>~1500) and number or processors was >60.
I never fully understood the nature of this bug, but figured out a
workaround. The issue was related to the first MPI_Send() call by
mpi_worker() and/or the first MPI_Probe() call by mpi_master(). It
seems like there is a 'time-out' that occurs by at least 1 worker that
causes the MPI process to fail when the initial mpi_master() step of
forecasting the total run time takes too long. It is only after this
forecast is complete that the master calls MPI_Probe() to receive the
worker's initial MPI_Send() (which is sent almost immediately after the
process starts). For large models, the forecast step can take up to
several minutes, so my guess is that the worker's MPI_Send() gives up
after too long. (If I simplify the forecast so it goes quicker then
the failure does not occur.) However, I never understood why this only
occurs with many processors, as I've never observed the failure with
less than 60 processors. 

The bug is logged in BUGTRAX, but not checked for in the testsuite,
since it requires MPI and only occurs for large jobs that commonly
take on the order of hours.

Updated 1.1rc2 release notes to mention that this bug has been fixed. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4338 db3e70e4-12e1-0310-b23f-81a6e3d09ba7