rmark-rocplot.c: Fixed serious bug: a key value was not being initialized to zero. The 'false_pos_per_q' variable (false positives per query) in the make_plot() function was not being properly initialized to zero.
This bug doesn't change any of my results on login-eddy, because the
value just happened to start out very near zero (< 10^-314) on the
specific compiled executable I used to get benchmark results. However,
when I run it on my laptop the bug changed the output relative to
login-eddy. (The changed results on my laptop reverted to the correct
ones if --enable-debugging was used with configure.)

00README and rmark-po-pp.sh: 
Fixed incorrect description of recently
updated usage of rmark-pp.sh, and updated rmark-po-pp.sh to be
consistent with rmark-pp.sh



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4487 db3e70e4-12e1-0310-b23f-81a6e3d09ba7