Logged and fixed bug i41 (first bug found since 1.1 release in October 2013), a minor bug related to cmconvert -b (binary) output. In certain cases, including when called as part of an 'ssh' command, and when piped (e.g. 'cmconvert -b my.1p0.cm | cat'), 'cmconvert -b' will fail. Reported by Eric Deveaud at the Institut Pasteur.
The fix I implemented, which was recommended by Eric, was to add a
'-o' option to cmconvert and require it when '-b' is used. 

No script to test this bug was added to the testsuite because once the
'-o' was added and required with '-b' it became impossible to
reproduce.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4664 db3e70e4-12e1-0310-b23f-81a6e3d09ba7