Implemented a better fix for bug i41 which was fixed in a recent revision (r4664).
The new fix is in cm_file.c::cm_file_WriteBinary() which tries to
determine the file offset at which the p7 filter is stored because
this can be returned optionally to the caller. When the binary output
is piped to cat or used in combination with an 'ssh' command this
offset cannot be determined (for reasons I don't pretend to
understand).  However, cmconvert's cm_file_WriteBinary() does not
request this offset be returned so the code doesn't need to try to
determine it.

So, the fix is to check if the file offset for the fp7 is going to be
returned before trying to determine it. Since it's not in cmconvert,
we avoid trying to determine it and avoid the failure.

I removed the requirement of '-o' with '-b' in cmconvert, which was
added as part of the r4664 fix. I left in the '-o' option in though,
since it could be useful.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4666 db3e70e4-12e1-0310-b23f-81a6e3d09ba7