Added the first new post-v1.1 feature that's not a bug fix. The maximum Hmm banded matrix size that's allowed in a search/scan pipeline is now model-size-dependent. Previously it was 128.0 Mb for all models by default (but settable with --mxsize). It's still settable with --mxsize, but by default it is now scaled from 128.0 Mb to 512.0 Mb depending on the current model's window size (W) parameter. For models with W<=1000 (the vast majority of Rfam models) it will be 128.0. For models with W>=3000 it will be 512 Mb. For anything in between (1000 < W < 3000), the maximum size is calculated as a linear function of W capped at 128 and 512.
The tradeoff here is, as usual, speed vs sensitivity. A smaller
maximum size results in slightly faster searches/scans because the HMM
banded matrix size limit acts as a filter, with higher stringency for
smaller maximum matrix sizes. But in my testing for LSU models, 128.0
is too limiting and causes the loss of an unacceptable number of real
LSU hits. Using 128.0 Mb is also probably a bit too strict for SSU
models, for which the new scheme sets the matrix size to about 300
Mb. 


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1@4678 db3e70e4-12e1-0310-b23f-81a6e3d09ba7