Make the build reproducible
Whilst working on the Reproducible Builds effort [0] we noticed that
infernal could not be built reproducibly.

This is because it embeds the current build timestamp and the absolute
build directory from which it was built into generated files. This
patch checks that if SOURCE_DATE_EPOCH [1] is set and, if so, omits
these fields.

This was originally filed in Debian as #946315 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/946315

Signed-off-by: Chris Lamb <lamby@debian.org>