Adds easel/formats symlink to hmmer/
This symlink is necessary to allow two hmmer testsuite tests that use files in easel/formats/ to pass when built as part of infernal and when the build directory and src directory are different. It may be easier in the future to just change the hmmer tests to not rely on files in easel/formats (by copying those files to hmmer/testsuite/) but for the infernal 1.1.5 release it is too late to make this change in hmmer because hmmer 3.4 was just released. I plan to make a hmmer PR after the 1.1.5 release to suggest the alternative.