Fixes cmscan thread race, in shared hmmfile errbuf
cmscan's master thread reads MSV profiles using p7_oprofile_ReadMSV()
 which uses P7_HMMFILE's errbuf. cmscan worker threads also (previously) used the same errbuf
(in cmfp->hfp->errbuf) if they progressed to the glocal envelope definition stage of
 cm_Pipeline() which requires the worker to read the full (general, non-optimized) HMM
from the HMM file which it needs because it configures it for glocal alignment which is
not possible with an optimized profile. The fix is to have infernal's read_{asc,bin}30hmm
functions (which are copies of the same functions from hmmer) to use the hfp->rr_errbuf
recently added by Sean to hmmer (commit 3867dac (related to hmmer github issues #277 and #278).
This works because previously neither cmscan master nor worker threads used rr_errbuf.