Fixes cmscan thread race related to CM_FILE errbuf
In cmscan, the master thread modifies cmfp->errbuf in cm_p7_oprofile_ReadMSV
and worker threads potentially modify cmfp->errbuf in cm_file_Read() if full CM is needed for
the final pipeline stages. Fix is to add a new msv_errbuf for exclusive use by cm_p7_oprofile_ReadMSV.
This is analogous to Sean's fix in hmmer to a similar thread race issue in hmmscan by the addition
of rr_errbuf to P7_HMMFILE (commit 3867dac, related to hmmer github issues #277 and #278).