Silences icx compiler warning
This fixes a mistake when initializing the cp9 HMM structure
of setting the begin and end probabilities from the 0th node to
to -INFTY (int). icx caught this as a implicit int to float
conversion. This should have been innocuous because the begin
and end probabilities from the 0th node are never used.