Silences gcc compiler snprintf warnings
Alignment displays string buffers are allocated carefully to have exactly the
number of necessary characters, which includes consensus positions in the annotation
of truncated begins/ends. This length (and corresponding number of chars) is computed
by the function, but gcc 12.2.0 on odyssey with --enable-debugging (-Wall) throws
a warning about strings potentially being truncated or out of range. I think this
is because the compiler doesn't realize the integer lengths are bounded (and can't
go as high as INTMAX for example). The way this commit silences the warnings is to
explicitly check the snprintf return value, and fail if it is not the expected value.